Devided into two servers
