+++ auth.html
@@ -9,7 +9,7 @@
 </head>
 <body>
     <h1>authentication page</h1>
-    <p><a href="./form">to form page</a></p>
-    <p><a href="./out">to reset cookie</a></p>
+    <p><a href="http://127.0.0.1:8001/form">to form page</a></p>
+    <p><a href="http://127.0.0.2:8002/out">to reset cookie</a></p>
 </body>
 </html>
