+++ charge.html
@@ -11,8 +11,8 @@
     <h1>charge page</h1>
     <p>{{from_form.purchase}} $ were sent</p>
     <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Explicabo, quibusdam iste. Magnam ad quo adipisci veniam, error saepe? Impedit eum consectetur commodi ipsam quibusdam consequuntur beatae aut nulla blanditiis vel?</p>
-    <p><a href="./auth">to auth page</a></p>
-    <p><a href="./form">to form page</a></p>
-    <p><a href="./out">to reset cookie</a></p>
+    <p><a href="http://127.0.0.2:8002/auth">to auth page</a></p>
+    <p><a href="http://127.0.0.1:8001/form">to form page</a></p>
+    <p><a href="http://127.0.0.2:8002/out">to reset cookie</a></p>
 </body>
 </html>
