+++ form.html
@@ -9,10 +9,10 @@
 </head>
 <body>
     <h1>form page</h1>
-    <p><a href="./auth">auth_page</a></p>
-    <p><a href="./out">to reset cookie</a></p>
+    <p><a href="http://127.0.0.2:8002/auth">auth_page</a></p>
+    <p><a href="http://127.0.0.2:8002/out">to reset cookie</a></p>
     <h2>Please, fill out the form</h2>
-    <form method="POST" action="./charge">
+    <form method="POST" action="http://127.0.0.2:8002/charge">
         <input type="text" name="purchase" value=10>
         <p><input type="submit" name="Submit"></p>
     </form>
